craft-application migration cleanup: pack command (#1704)
This pack command existed for comparison with the new one, but is no
longer necessary. It got missed when I was removing some other
no-longer-used commands.
